+++ core/src/main/java/org/yes/cart/bulkjob/shoppingcart/BulkEmptyAnonymousShoppingCartProcessorImpl.java
@@ -85,7 +85,7 @@ public class BulkEmptyAnonymousShoppingCartProcessorImpl implements Runnable {
                 log.debug("Removed empty anonymous cart for {}, guid {}", scs.getCustomerEmail(), guid);
 
                 final CustomerOrder tempOrder = this.customerOrderService.findByReference(guid);
-                if (CustomerOrder.ORDER_STATUS_NONE.equals(tempOrder.getOrderStatus())) {
+                if (tempOrder != null && CustomerOrder.ORDER_STATUS_NONE.equals(tempOrder.getOrderStatus())) {
                     log.debug("Removing temporary order for cart guid {}", guid);
                     this.customerOrderService.delete(tempOrder);
                     removedOrders++;
